// frontend/plugins/welcome/src/api/models/EntPatient.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SUT SA Example API Patient
 * This is a sample server for SUT SE 2563
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    EntPatientEdges,
    EntPatientEdgesFromJSON,
    EntPatientEdgesFromJSONTyped,
    EntPatientEdgesToJSON,
} from './';

/**
 * 
 * @export
 * @interface EntPatient
 */
export interface EntPatient {
    /**
     * Address holds the value of the "Address" field.
     * @type {string}
     * @memberof EntPatient
     */
    address?: string;
    /**
     * Age holds the value of the "Age" field.
     * @type {number}
     * @memberof EntPatient
     */
    age?: number;
    /**
     * Allergic holds the value of the "Allergic" field.
     * @type {string}
     * @memberof EntPatient
     */
    allergic?: string;
    /**
     * Birthday holds the value of the "Birthday" field.
     * @type {string}
     * @memberof EntPatient
     */
    birthday?: string;
    /**
     * CardID holds the value of the "Card_id" field.
     * @type {string}
     * @memberof EntPatient
     */
    cardId?: string;
    /**
     * FirstName holds the value of the "First_name" field.
     * @type {string}
     * @memberof EntPatient
     */
    firstName?: string;
    /**
     * LastName holds the value of the "Last_name" field.
     * @type {string}
     * @memberof EntPatient
     */
    lastName?: string;
    /**
     * 
     * @type {EntPatientEdges}
     * @memberof EntPatient
     */
    edges?: EntPatientEdges;
    /**
     * ID of the ent.
     * @type {number}
     * @memberof EntPatient
     */
    id?: number;
}

export function EntPatientFromJSON(json: any): EntPatient {
    return EntPatientFromJSONTyped(json, false);
}

export function EntPatientFromJSONTyped(json: any, ignoreDiscriminator: boolean): EntPatient {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'address': !exists(json, 'Address') ? undefined : json['Address'],
        'age': !exists(json, 'Age') ? undefined : json['Age'],
        'allergic': !exists(json, 'Allergic') ? undefined : json['Allergic'],
        'birthday': !exists(json, 'Birthday') ? undefined : json['Birthday'],
        'cardId': !exists(json, 'Card_id') ? undefined : json['Card_id'],
        'firstName': !exists(json, 'First_name') ? undefined : json['First_name'],
        'lastName': !exists(json, 'Last_name') ? undefined : json['Last_name'],
        'edges': !exists(json, 'edges') ? undefined : EntPatientEdgesFromJSON(json['edges']),
        'id': !exists(json, 'id') ? undefined : json['id'],
    };
}

export function EntPatientToJSON(value?: EntPatient | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'Address': value.address,
        'Age': value.age,
        'Allergic': value.allergic,
        'Birthday': value.birthday,
        'Card_id': value.cardId,
        'First_name': value.firstName,
        'Last_name': value.lastName,
        'edges': EntPatientEdgesToJSON(value.edges),
        'id': value.id,
    };
}